NFC: Move patches to a separate file
